pyproject.toml
[project]
name = "awesome-quant"
version = "0.1.0"
description = ""
authors = [{ name = "Wilson Freitas", email = "wilson.freitas@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.2.0",
    "pandas>=2.2.0",
    "mypy>=1.14.0",
]
